requirements.txt
requests
websockets
asyncio
python-dotenv
